netApi

+++ src/ru/inventos/yum/DeliveryPriceReceiver.java
@@ -1,5 +1,5 @@
 package ru.inventos.yum;
 
 public interface DeliveryPriceReceiver {
-	public void receiveDeliveryPrice(float price, float freePrice);
+	public void receiveDeliveryPrice(int price, int freePrice);
 }
